change t_buffer.n type (int -> long)

+++ get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: susumuyagi <susumuyagi@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/25 14:04:42 by susumuyagi        #+#    #+#             */
-/*   Updated: 2023/05/30 14:06:24 by susumuyagi       ###   ########.fr       */
+/*   Updated: 2023/06/07 11:40:29 by susumuyagi       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@
 
 typedef struct s_buffer
 {
-	int		n;
+	long	n;
 	char	*bufp;
 	char	buf[BUFFER_SIZE];
 }			t_buffer;
